requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
xgboost
hyperopt
fastapi
uvicorn
python-dotenv
# -e .
